+++ linjiashop-admin-api/src/main/java/cn/enilu/flash/api/controller/system/MenuController.java
@@ -1,6 +1,5 @@
 package cn.enilu.flash.api.controller.system;
 
-import cn.enilu.flash.bean.constant.state.MenuStatus;
 import cn.enilu.flash.bean.core.AuthorizationUser;
 import cn.enilu.flash.bean.core.BussinessLog;
 import cn.enilu.flash.bean.entity.system.Menu;
@@ -8,10 +7,7 @@ import cn.enilu.flash.bean.enumeration.Permission;
 import cn.enilu.flash.bean.exception.ApplicationException;
 import cn.enilu.flash.bean.exception.ApplicationExceptionEnum;
 import cn.enilu.flash.bean.vo.front.Rets;
-import cn.enilu.flash.bean.vo.node.MenuNode;
-import cn.enilu.flash.bean.vo.node.Node;
-import cn.enilu.flash.bean.vo.node.RouterMenu;
-import cn.enilu.flash.bean.vo.node.ZTreeNode;
+import cn.enilu.flash.bean.vo.node.*;
 import cn.enilu.flash.cache.TokenCache;
 import cn.enilu.flash.service.system.LogObjectHolder;
 import cn.enilu.flash.service.system.MenuService;
@@ -41,6 +37,7 @@ import java.util.Map;
 @RequestMapping("/menu")
 public class MenuController extends BaseController {
 
+
     private Logger logger = LoggerFactory.getLogger(MenuController.class);
     @Autowired
     private MenuService menuService;
@@ -54,11 +51,36 @@ public class MenuController extends BaseController {
         List<RouterMenu> list = menuService.getSideBarMenus(shiroUser.getRoleList());
         return Rets.success(list);
     }
+
     @RequestMapping(value = "/list", method = RequestMethod.GET)
-    @RequiresPermissions(value = {Permission.MENU})
     public Object list() {
         List<MenuNode> list = menuService.getMenus();
         return Rets.success(list);
+    }
+
+    @RequestMapping(value = "/tree", method = RequestMethod.GET)
+    public Object tree() {
+        List<MenuNode> list = menuService.getMenus();
+        List<TreeSelectNode> treeSelectNodes = Lists.newArrayList();
+        for (MenuNode menuNode : list) {
+            TreeSelectNode tsn = transfer(menuNode);
+            treeSelectNodes.add(tsn);
+        }
+        return Rets.success(treeSelectNodes);
+    }
+
+    public TreeSelectNode transfer(MenuNode node) {
+        TreeSelectNode tsn = new TreeSelectNode();
+        tsn.setId(node.getCode());
+        tsn.setLabel(node.getName());
+        if (node.getChildren() != null && !node.getChildren().isEmpty()) {
+            List<TreeSelectNode> children = Lists.newArrayList();
+            for (MenuNode child : node.getChildren()) {
+                children.add(transfer(child));
+            }
+            tsn.setChildren(children);
+        }
+        return tsn;
     }
 
     @RequestMapping(method = RequestMethod.POST)
@@ -66,21 +88,18 @@ public class MenuController extends BaseController {
     @RequiresPermissions(value = {Permission.MENU_EDIT})
     public Object save(@ModelAttribute @Valid Menu menu) {
         //判断是否存在该编号
-        if(menu.getId()==null) {
+        if (menu.getId() == null) {
             String existedMenuName = ConstantFactory.me().getMenuNameByCode(menu.getCode());
             if (StringUtil.isNotEmpty(existedMenuName)) {
                 throw new ApplicationException(ApplicationExceptionEnum.EXISTED_THE_MENU);
             }
-            menu.setStatus(MenuStatus.ENABLE.getCode());
         }
 
         //设置父级菜单编号
         menuService.menuSetPcode(menu);
-        if(menu.getId()==null){
+        if (menu.getId() == null) {
             menuService.insert(menu);
-        }else {
-            Menu old = menuService.get(menu.getId());
-            LogObjectHolder.me().set(old);
+        } else {
             menuService.update(menu);
         }
         return Rets.success();
@@ -95,7 +114,7 @@ public class MenuController extends BaseController {
             throw new ApplicationException(ApplicationExceptionEnum.REQUEST_NULL);
         }
         //演示环境不允许删除初始化的菜单
-        if(id.intValue()<70){
+        if (id.intValue() < 70) {
             return Rets.failure("演示环境不允许删除初始菜单");
         }
         //缓存菜单的名称
@@ -112,7 +131,7 @@ public class MenuController extends BaseController {
     public Object menuTreeListByRoleId(Integer roleId) {
         List<Long> menuIds = menuService.getMenuIdsByRoleId(roleId);
         List<ZTreeNode> roleTreeList = null;
-        if (menuIds==null||menuIds.isEmpty()) {
+        if (menuIds == null || menuIds.isEmpty()) {
             roleTreeList = menuService.menuTreeList(null);
         } else {
             roleTreeList = menuService.menuTreeList(menuIds);
@@ -121,10 +140,10 @@ public class MenuController extends BaseController {
         List<Node> list = menuService.generateMenuTreeForRole(roleTreeList);
 
         //element-ui中tree控件中如果选中父节点会默认选中所有子节点，所以这里将所有非叶子节点去掉
-        Map<Long,ZTreeNode> map = cn.enilu.flash.utils.Lists.toMap(roleTreeList,"id");
-        Map<Long,List<ZTreeNode>> group = cn.enilu.flash.utils.Lists.group(roleTreeList,"pId");
-        for(Map.Entry<Long,List<ZTreeNode>> entry:group.entrySet()){
-            if(entry.getValue().size()>1){
+        Map<Long, ZTreeNode> map = cn.enilu.flash.utils.Lists.toMap(roleTreeList, "id");
+        Map<Long, List<ZTreeNode>> group = cn.enilu.flash.utils.Lists.group(roleTreeList, "pId");
+        for (Map.Entry<Long, List<ZTreeNode>> entry : group.entrySet()) {
+            if (entry.getValue().size() > 1) {
                 roleTreeList.remove(map.get(entry.getKey()));
             }
         }
@@ -132,7 +151,7 @@ public class MenuController extends BaseController {
         List<Long> checkedIds = Lists.newArrayList();
         for (ZTreeNode zTreeNode : roleTreeList) {
             if (zTreeNode.getChecked() != null && zTreeNode.getChecked()
-                    &&zTreeNode.getpId().intValue()!=0) {
+                    && zTreeNode.getpId().intValue() != 0) {
                 checkedIds.add(zTreeNode.getId());
             }
         }